+++ app/build.gradle
@@ -37,8 +37,8 @@ dependencies {
 
     compile 'com.android.support:recyclerview-v7'
 
-    compile 'com.android.support:preference-v14:25.3.1'
 
+    compile 'com.android.support:preference-v14:25.3.1'
     compile 'com.android.support:appcompat-v7:25.3.1'
     compile 'com.android.support.constraint:constraint-layout:1.0.2'
     compile 'com.android.support:design:25.3.1'
