+++ app/src/main/java/com/kevguev/mobile/vidly/ui/FavoritesFragment.java
@@ -1,0 +1,34 @@
+package com.kevguev.mobile.vidly.ui;
+
+import android.os.Bundle;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.kevguev.mobile.vidly.R;
+
+/**
+ * Created by Kevin Guevara on 5/24/2017.
+ */
+
+public class FavoritesFragment extends Fragment {
+
+    public FavoritesFragment(){
+
+    }
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_main_list, container, false);
+    }
+
+}
