+++ app/src/main/java/com/kevguev/mobile/vidly/ui/MapFragment.java
@@ -1,0 +1,34 @@
+package com.kevguev.mobile.vidly.ui;
+
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.support.v4.app.Fragment;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.kevguev.mobile.vidly.R;
+
+/**
+ * Created by Kevin Guevara on 5/24/2017.
+ */
+
+public class MapFragment extends Fragment {
+
+    public MapFragment(){
+
+    }
+
+    @Override
+    public void onCreate(@Nullable Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+    }
+
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_main_list, container, false);
+
+    }
+}
