+++ app/src/main/res/values/strings.xml
@@ -4,6 +4,7 @@
     <string name="pref_search_query">pref_search_query</string>
     <string name="pref_location">pref_location</string>
     <string name="pref_radius">pref_radius</string>
+    <string name="pref_published_after">pref_published_after</string>
     <!-- TODO: Remove or change this placeholder text -->
     <string name="hello_blank_fragment">Hello blank fragment</string>
 </resources>
